+++ src/routes/_required-header/timetable/index.tsx
@@ -1,15 +1,8 @@
 import { createFileRoute } from '@tanstack/react-router';
-import TimeTable from '../../../components/timetable/TimeTable';
+import loadFailure from '../../../api/loadFailure';
 
 export const Route = createFileRoute('/_required-header/timetable/')({
-  component: () => <TimeTablePage />,
+  loader: async () => await loadFailure(),
+  pendingComponent: () => <p>로딩중</p>,
+  errorComponent: ({ error }) => <p>{error.message}</p>,
 });
-
-function TimeTablePage() {
-  return (
-    <main className="pb-[20px]">
-      <h1 className="text-center text-[36px] my-[50px]">MARCH 27</h1>
-      <TimeTable />
-    </main>
-  );
-}
